fix(quan-ly-quan): don ma lam ket ban, thanh toan hai lan, va ba trang thai quan
Dot 2 cua ke hoach hoan thien — mang Van hanh quan (M4). Bon nhom viec, gop lam
mot commit vi chung dung chung tep (OrderController, trang Ban hang).

1. DON MA LAM KET BAN. Don duoc ghi truoc roi moi them tung dong mon; dong nao
   khong hop le thi 422 bay ra nhung cai don rong da nam lai o trang thai dang
   phuc vu — ban ket tren so do, va lan goi mon sau bi nhanh "chong tao order
   trung" tra ve chinh cai don ma do. O update() con nang hon: cac dong cu bi xoa
   ngay dau vong lap nen mot request hong lam mat luon mon cua ban dang ngoi.
   -> Tach chuanBiDong() (kiem + tinh, khong ghi) khoi ghiDong(). Kiem het moi
      dong truoc; qua duoc het moi bat dau ghi. MongoDB standalone khong co
      transaction nen moi thu da ghi la o lai vinh vien.

2. THANH TOAN HAI LAN. Kiem status o dau ham khong du: hai request song song
   (bam dup, hai tab) deu doc thay 'active' roi cung di tiep, ra HAI ma phieu.
   -> Chot don bang mot phep ghi co dieu kien (where status = active). Ap dung ca
      cho cancel() de chan canh vua thu tien vua ghi so la da huy.
   Kem theo: don khong con mon nao thi khong thanh toan duoc (hoa don 0d van
   duoc dem la mot don trong so), va go not dong cuoi o POS = HUY don chu khong
   luu don rong.

3. BA TRANG THAI QUAN co y nghia that. Truoc day chi la ba cai nhan mau: chon
   "Ngung hoat dong" xong van ban hang binh thuong. Nay:
   - Da dong cua: khong mo don moi, nhung ban dang ngoi van goi them va thanh
     toan duoc (khong nhot tien cua khach), thuc don van sua duoc.
   - Ngung hoat dong: chi tra cuu, khong ban va khong sua thuc don.
   Y nghia nay duoc viet ra ngay tren trang Thong tin quan va nhac lai o POS.

4. LUAT THUC DON. Gia topping luu thanh so nguyen dong nhu gia mon (cast 'float'
   cu lam topping thanh cho duy nhat trong thuc don tra ve so thuc, roi nhan qua
   so phan va so ly thanh so le tren hoa don). Khong hai cai ban / hai danh muc
   cung ten trong mot quan. An danh muc = an ca mon ben trong khoi POS, chan ca o
   may chu.

47 bai kiem thu may chu moi (44 -> 91 bai, 203 khang dinh) va 9 bai frontend cho
quy tac cong don / tach dong cua gio hang.

+++ backend/app/Models/Topping.php
@@ -12,7 +12,11 @@ class Topping extends Model
 
     protected $casts = [
         'is_available' => 'boolean',
-        'price' => 'float',
+        // Số NGUYÊN như base_price của Item và price của ItemPrice. Trước đây là
+        // 'float', nên topping là chỗ duy nhất trong thực đơn trả về số thực: giá
+        // đó được nhân với số phần rồi nhân với số ly ở OrderController, và hóa đơn
+        // ra số lẻ mà không ai gõ vào. Tiền Việt không có đơn vị nhỏ hơn đồng.
+        'price' => 'integer',
     ];
 
     public function cafe()
